requirements.txt
cryptography>=3.4.8
pyOpenSSL>=21.0.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
websockets>=13.1
asyncio-mqtt
python-multipart
pytapo
dotenv
tapo
aiofiles==24.1.0
aiosqlite==0.21.0
beautifulsoup4==4.13.4
brotli==1.1.0
cachetools==5.5.2
chardet==5.2.0
crawl4ai==0.6.3
cssselect==1.3.0
dataclasses-json==0.6.7
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
filelock==3.18.0
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanize==4.12.3
importlib-metadata==8.7.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-community==2.0.7
langchain-text-splitters==0.3.8
langsmith==0.3.45
litellm==1.72.4
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mypy-extensions==1.1.0
nltk==3.9.1
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
packaging==24.2
pillow==10.4.0
playwright==1.52.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic-settings==2.9.1
pyee==13.0.0
pygments==2.19.1
pyparsing==3.2.3
pyperclip==1.9.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
snowballstemmer==2.2.0
soupsieve==2.7
sqlalchemy==2.0.41
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspect==0.9.0
uritemplate==4.2.0
xxhash==3.5.0
zipp==3.23.0
zstandard==0.23.0